requirements.txt
South==0.7.6
django-modeltranslation
django-wysiwyg-redactor